sda.enkf

State Variable Data Assimilation: Ensemble Kalman Filter

State Variable Data Assimilation: Ensemble Kalman Filter and Generalized ensemble filter

sda.enkf.original(settings, obs.mean, obs.cov, IC = NULL, Q = NULL,
  adjustment = TRUE, restart = NULL)

sda.enkf(settings, obs.mean, obs.cov, Q = NULL, restart = F,
  control = list(trace = T, interactivePlot = T, TimeseriesPlot = T,
  BiasPlot = F, plot.title = NULL, debug = FALSE), ...)

Arguments

settings
PEcAn settings object

obs.mean
list of observations of the means of state variable (time X nstate)

obs.cov
list of observations of covariance matrices of state variables (time X nstate X nstate)

IC
initial conditions

Q
process covariance matrix given if there is no data to estimate it

adjustment
flag for using ensemble adjustment filter or not

restart
Used for iterative updating previous forecasts. This is a list that includes ens.inputs, the list of inputs by ensemble member, params, the parameters, and old_outdir, the output directory from the previous workflow. These three things are needed to ensure that if a new workflow is started that ensemble members keep there run-specific met and params. See Details

control
List of flags controlling the behaviour of the SDA. trace for reporting back the SDA outcomes, interactivePlot for plotting the outcomes after each step, TimeseriesPlot for post analysis examination, BiasPlot for plotting the correlation between state variables, plot.title is the title of post analysis plots and debug mode allows for pausing the code and examinign the variables inside the function.

settings
PEcAn settings object

obs.mean
List of dataframe of observation means, named with observation datetime.

obs.cov
List of covariance matrices of state variables , named with observation datetime.

Q
Process covariance matrix given if there is no data to estimate it.

restart
Used for iterative updating previous forecasts. When the restart is TRUE it read the object in SDA folder written from previous SDA.

Value

NONE

NONE